Modifica o component Card
Diferente do HomeCard, esse component Card.jsx será responsável pelas páginas de gêneros de filmes e pela SearchPage. A grande diferença de um para outro é que na HomePage os cards serão mostrados em carousel e nas outras páginas que citei anteriormente, todos os cards terão um 'wrap', ou seja, após preencher o espaço, eles cairão para linha de baixo
